Add gateway and gateway connections entities
And create apiController
