+++ module/Application/src/Exceptions/AuthenticationException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Application\Exceptions;
+
+class AuthenticationException extends ThingsException
+{
+}
