+++ module/Application/src/Exceptions/InvalidPayloadException.php
@@ -1,8 +1,6 @@
 <?php
 namespace Application\Exceptions;
 
-use Exception;
-
-class InvalidPayloadException extends Exception
+class InvalidPayloadException extends ThingsException
 {
 }
