+++ module/Application/src/Exceptions/ThingsException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Application\Exceptions;
+
+use Exception;
+
+class ThingsException extends Exception
+{
+}
